+++ .dialyzer_ignore.exs
@@ -1,0 +1,4 @@
+[
+  # Mix.Task callback info not available in Dialyzer PLT
+  {"lib/mix/tasks/mpp.manifest.ex", :callback_info_missing}
+]